requirements.txt
streamlit
python-dotenv
openai==1.23.2
streamlit-authenticator
deta